requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==43.0.1
distlib==0.3.8
Django==4.2.14
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.9.2
fakeredis==2.23.3
filelock==3.15.4
gunicorn==22.0.0
identify==2.6.0
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.9.1
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pytest==8.2.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-order==1.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.0.7
requests==2.32.3
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
